+++ docs/RELEASE_NOTES.md
@@ -1,3 +1,3 @@
-## Improvements
+## Issues
 
-- Added new convention to set options using env variables
+- Fixed an issue where inherited options were not displayed.